// src/types/language/image-optimizer.ts
export interface ImageOptimizerTranslations {
  // Image optimizer
  imageOptimizer: string;
  upload: string;
  uploadImage: string;
  basic: string;
  filters: string;
  brightness: string;
  contrast: string;
  saturation: string;
  sharpness: string;
  reset: string;
  optimizeWithAI: string;
  optimizing: string;
  apply: string;
  normal: string;
  imageOptimized: string;
  imageOptimizedDesc: string;
}